% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_cite.R
\name{mod_cite_server}
\alias{mod_cite_server}
\title{Citation server-side function}
\usage{
mod_cite_server(id)
}
\arguments{
\item{id}{Internal parameter for Shiny.}
}
\description{
Citation server-side function
}
