% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_products_server.R
\name{mod_products_server}
\alias{mod_products_server}
\title{Product profile server-side function}
\usage{
mod_products_server(id)
}
\arguments{
\item{id}{Internal parameter for Shiny.}
}
\description{
A shiny Module.
}
