% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_server.R
\name{od_treemap}
\alias{od_treemap}
\title{Origin-Destination Treemap}
\usage{
od_treemap(d, d2)
}
\arguments{
\item{d}{input dataset for values}

\item{d2}{input dataset for colours}
}
\description{
Origin-Destination Treemap
}
