% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_server.R
\name{p_treemap}
\alias{p_treemap}
\title{Product Treemap}
\usage{
p_treemap(d, d2, title = NULL)
}
\arguments{
\item{d}{input dataset for values}

\item{d2}{input dataset for colours}

\item{title}{title for the treemap}
}
\description{
Product Treemap
}
